createRootUIViewControllerState
